+++ src/ws_server/types.ts
@@ -1,0 +1,8 @@
+export interface Player {
+  name: string;
+  password: string;
+}
+
+export interface Room {
+  [key: string]: Player[];
+}
